requirements.txt
bcrypt==3.1.3
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.0
Flask-uWSGI-WebSocket==0.6.0
geographiclib==1.48
python-resize-image==1.1.11
pyresttest==1.7.1
